chore(appeals): moving string case utils out to top level utils from web and api (no-ticket)
