+++ appeals/api/src/server/endpoints/lpa-questionnaires/lpa-questionnaires.controller.js
@@ -8,9 +8,9 @@ import { allLpaQuestionnaireOutcomesAreComplete } from '#utils/is-awaiting-linke
 import { isParentAppeal } from '#utils/is-linked-appeal.js';
 import logger from '#utils/logger.js';
 import stringTokenReplacement from '#utils/string-token-replacement.js';
-import { camelToScreamingSnake, capitalizeFirstLetter } from '#utils/string-utils.js';
 import { APPEAL_TYPE } from '@pins/appeals/constants/common.js';
 import * as CONSTANTS from '@pins/appeals/constants/support.js';
+import { camelToScreamingSnake, capitalizeFirstLetter } from '@pins/appeals/utils/string-case.js';
 import {
 	updateLPAQuestionnaire,
 	updateLPAQuestionnaireValidationOutcome
