+++ appeals/web/src/server/appeals/appeal-details/internal-correspondence/internal-correspondence.controller.js
@@ -19,7 +19,7 @@ import { getDocumentFileType } from '#appeals/appeal-documents/appeal.documents.
 import logger from '#lib/logger.js';
 import { mapFolderNameToDisplayLabel } from '#lib/mappers/utils/documents-and-folders.js';
 import { addNotificationBannerToSession } from '#lib/session-utilities.js';
-import { capitalizeFirstLetter } from '#lib/string-utilities.js';
+import { capitalizeFirstLetter } from '@pins/appeals/utils/string-case.js';
 import { capitalize } from 'lodash-es';
 import { documentNameFromCategory } from './internal-correspondence.service.js';
