+++ appeals/web/src/server/lib/appeal-status.js
@@ -1,5 +1,5 @@
 import { isAppealTypeEnabled } from '#common/feature-flags-appeal-types.js';
-import { capitalizeFirstLetter } from '#lib/string-utilities.js';
+import { capitalizeFirstLetter } from '@pins/appeals/utils/string-case.js';
 
 /** @typedef {import('#appeals/appeal-details/appeal-details.types.js').WebAppeal} WebAppeal */
 /** @typedef {import('#appeals/personal-list/personal-list.mapper').PersonalListAppeal} PersonalListAppeal */
