+++ appeals/web/src/server/lib/mappers/data/appellant-case/submappers/application-type.js
@@ -1,5 +1,5 @@
 import { textSummaryListItem } from '#lib/mappers/components/index.js';
-import { capitalizeFirstLetter } from '#lib/string-utilities.js';
+import { capitalizeFirstLetter } from '@pins/appeals/utils/string-case.js';
 import { APPEAL_TYPE_OF_PLANNING_APPLICATION } from '@planning-inspectorate/data-model';
 
 /**
