+++ appeals/web/src/server/lib/mappers/data/appellant-case/submappers/development-type.js
@@ -1,6 +1,6 @@
 import { textSummaryListItem } from '#lib/mappers/components/index.js';
-import { capitalizeFirstLetter } from '#lib/string-utilities.js';
 import { APPEAL_DEVELOPMENT_TYPES } from '@pins/appeals/constants/appellant-cases.constants.js';
+import { capitalizeFirstLetter } from '@pins/appeals/utils/string-case.js';
 
 /** @type {import('../mapper.js').SubMapper} */
 export const mapDevelopmentType = ({ appellantCaseData, currentRoute, userHasUpdateCase }) => {
