+++ appeals/web/src/server/lib/validators/date-input.validator.js
@@ -1,5 +1,5 @@
 import { MONTH_SET } from '#lib/constants.js';
-import { toCamelCase } from '#lib/string-utilities.js';
+import { toCamelCase } from '@pins/appeals/utils/string-case.js';
 import { createValidator } from '@pins/express';
 import { body } from 'express-validator';
 import { lowerCase } from 'lodash-es';
